Proof of concept
This is a CA for the etcd environment in a container.
Sprinkle in some duplicity a la git-deploy and it could even be HA.
I'd spec-ed this to wrap the openssl CLI, but wrapping coreos's etcd-ca
tool is even easier path to demo.

We want the CA to be available for API calls so that etcd servers can
come/go from an AutoScalingGroup and always get a certificate that
matches their current IP. They'll use it for peering _and_ client
connections (clients can also request certificates from
etcd-cert-server).

Per usual, make+Vagrant are used to drive a test environment:
etcd-cert-server runs on a "core-master" node, and core-01, core-02 and
core-03 nodes acquire certificates during bootstrap.

Split server/peer keys, add client keys