Proof of concept
This is a CA for the etcd environment in a container.
Sprinkle in some duplicity a la git-deploy and it could even be HA.
I'd spec-ed this to wrap the openssl CLI, but wrapping coreos's etcd-ca
tool is even easier path to demo.

We want the CA to be available for API calls so that etcd servers can
come/go from an AutoScalingGroup and always get a certificate that
matches their current IP. They'll use it for peering _and_ client
connections (clients can also request certificates from
etcd-cert-server).

Per usual, make+Vagrant are used to drive a test environment:
etcd-cert-server runs on a "core-master" node, and core-01, core-02 and
core-03 nodes acquire certificates during bootstrap.

There are a few layers of hardenning planned:
- etcd-ca stores _every_ keypair, openssl CLI could generate + sign keys
during the request cycle to avoid this.
- etcd-cert-server port remains open to instances after they've received
certificates, in this implementation subsequent requests leak the active
private key (ruh oh).
I think this can be mitigated by specifying 2 SG's on the etcd ASG:
a base and an extra that only allows connections to etcd-cert-server.
This would allow etcd-cert-server to respond to a fresh instance's
request, then remove the "extra" SG from that instance to prevent it
from connecting again.