/v1/certs/finalize, this port will self destruct
This uses the EC2 API to look up the instance that made a request, then
remove it from a specific security group.
By setting that security group to one that allows the initial connection
to be made, the /finalize resource serves to prevent any further
connections from that host.

This can be used by clients to fetch etcd certificates, then "burn the
bridge" to prevent any subsequent certificates being issued, mitigating
some risk from the "always-on-auto-CA" design.